Rendering Reorganization
I've reorganized rendering during my quest to separate the environment
and the visualization. At the moment it's completely broken and I'm too
tired to find it now, so don't expect visualization to work.
Secondly, I'm not happy with it as is, so I may get rid of the "Core"
files idea and just have a completely separate set of files for the
Environment and the Visualizer.
Oh well.


Former-commit-id: 5dcbb903e1fa623704285037ad452f547a466e56 [formerly 76064cf9a2c0601ee0c70e94a608b45f5cacae0e]
Former-commit-id: e6db92f18db54da71349146b4514a94d0ba88d85
Former-commit-id: 35c3125951f04fcebbf4427d0f3db7fd4b03ff34