Took out old rendering formats.
Decreases size of executable, which is good, at the expense of a few old
replays we won't be able to read any more through the main version.
I've copied the old executable, however, to allow  the occasional
browsing of old games as one might wish, but we'll overall move away
from that.


Former-commit-id: 05d8312cfe654d396b74561f2a718a5596deca5c
Former-commit-id: 82ab1990c8992b8d6901d93c96f5a3f08dbb2c73