Command Line Functionality!
You can now set it as a default program for opening .hlt files!
However, to make this happen, you are going to need to put all of the
runtime resources into the folder of the program which you set as the
default. That includes:
freetype6.dll
zlib1.dll
shaders
settings
fonts
logs (optional but preferable)

I also took out ftgl, since we don't use that anymore.


Former-commit-id: 456b76e4a8b4d838784eb767858d617a66ecf148
Former-commit-id: 2189add7a14513e66dbb5dbdd676729eef0aed78