Update rules_game.php
just fixing grammar.  `excepting` -> `except`
When pieces from opposing players try to occupy either the same or cardinally adjacent sites (diagonals are not included), the battle will be resolved according to the relative strengths of the pieces, as each piece decreases the strength of every adjacent or coinciding opposing piece by its own strength. Pieces with a strength of 0 or less are removed from the game, `excepting`* pieces with a strength of 0 which have not engaged in combat during that turn.